Renamed internals to internal as per golang-standards/project-layout
