+++ internal/config/config.go
@@ -5,7 +5,7 @@ import (
 	"os"
 	"sync"
 
-	"github.com/mrlhansen/idrac_exporter/internals/logging"
+	"github.com/mrlhansen/idrac_exporter/internal/logging"
 	"gopkg.in/yaml.v2"
 )
