+++ internal/redfish/client.go
@@ -10,8 +10,8 @@ import (
 	"strings"
 	"time"
 
-	"github.com/mrlhansen/idrac_exporter/internals/config"
-	"github.com/mrlhansen/idrac_exporter/internals/logging"
+	"github.com/mrlhansen/idrac_exporter/internal/config"
+	"github.com/mrlhansen/idrac_exporter/internal/logging"
 )
 
 const redfishRootPath = "/redfish/v1"
