+++ internal/redfish/model.go
